子コンポーネントで完結
